ds_migration_app_dev/requirements.txt
alembic==1.4.3
blist==1.3.6
boto3==1.16.19
botocore==1.19.19
cassandra-driver==3.24.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-CQLAlchemy==1.2.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
geomet==0.2.1.post1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
Werkzeug==1.0.1
